pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "restorers"
dynamic = ["version"]
description = "Image Restoration toolkit in TensorFlow"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "matplotlib>=3.5.3",
    "ml-collections==0.1.1",
    "pillow",
    "tqdm",
    "wandb>=0.13.6"
]

[project.optional-dependencies]
dev=[
    "isort>=5.10.1",
    "black>=22.6.0"
]
test=[
    "pytest==7.1.2"
]
docs=[
    "mkdocs",
    "mkdocstrings[python]",
    "mkdocs-material>=9.1.5",
    "mkdocs-minify-plugin",
    "mkdocs-glightbox"
]

[project.urls]
homepage = "https://github.com/soumik12345/restorers"
repository = "https://github.com/soumik12345/restorers"
